requirements.txt
Django>=4.2
cryptography
gunicorn
dj-database-url
python-decouple
whitenoise
